ref: move performance integrations into folder
