+++ packages/node-experimental/src/integrations/tracing/graphql.ts
@@ -1,21 +1,22 @@
 import type { Instrumentation } from '@opentelemetry/instrumentation';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
-import { MongoDBInstrumentation } from '@opentelemetry/instrumentation-mongodb';
+import { GraphQLInstrumentation } from '@opentelemetry/instrumentation-graphql';
 import { defineIntegration } from '@sentry/core';
 import type { Integration, IntegrationFn } from '@sentry/types';
 
-import { addOriginToSpan } from '../utils/addOriginToSpan';
+import { addOriginToSpan } from '../../utils/addOriginToSpan';
 import { NodePerformanceIntegration } from './NodePerformanceIntegration';
 
-const _mongoIntegration = (() => {
+const _graphqlIntegration = (() => {
   return {
-    name: 'Mongo',
+    name: 'Graphql',
     setupOnce() {
       registerInstrumentations({
         instrumentations: [
-          new MongoDBInstrumentation({
+          new GraphQLInstrumentation({
+            ignoreTrivialResolveSpans: true,
             responseHook(span) {
-              addOriginToSpan(span, 'auto.db.otel.mongo');
+              addOriginToSpan(span, 'auto.graphql.otel.graphql');
             },
           }),
         ],
@@ -24,20 +25,20 @@ const _mongoIntegration = (() => {
   };
 }) satisfies IntegrationFn;
 
-export const mongoIntegration = defineIntegration(_mongoIntegration);
+export const graphqlIntegration = defineIntegration(_graphqlIntegration);
 
 /**
- * MongoDB integration
+ * GraphQL integration
  *
- * Capture tracing data for MongoDB.
+ * Capture tracing data for GraphQL.
  *
- * @deprecated Use `mongoIntegration()` instead.
+ * @deprecated Use `graphqlIntegration()` instead.
  */
-export class Mongo extends NodePerformanceIntegration<void> implements Integration {
+export class GraphQL extends NodePerformanceIntegration<void> implements Integration {
   /**
    * @inheritDoc
    */
-  public static id: string = 'Mongo';
+  public static id: string = 'GraphQL';
 
   /**
    * @inheritDoc
@@ -47,15 +48,16 @@ export class Mongo extends NodePerformanceIntegration<void> implements Integrati
   public constructor() {
     super();
     // eslint-disable-next-line deprecation/deprecation
-    this.name = Mongo.id;
+    this.name = GraphQL.id;
   }
 
   /** @inheritDoc */
   public setupInstrumentation(): void | Instrumentation[] {
     return [
-      new MongoDBInstrumentation({
+      new GraphQLInstrumentation({
+        ignoreTrivialResolveSpans: true,
         responseHook(span) {
-          addOriginToSpan(span, 'auto.db.otel.mongo');
+          addOriginToSpan(span, 'auto.graphql.otel.graphql');
         },
       }),
     ];
