+++ packages/node-experimental/src/integrations/tracing/mongo.ts
@@ -1,21 +1,21 @@
 import type { Instrumentation } from '@opentelemetry/instrumentation';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
-import { MySQL2Instrumentation } from '@opentelemetry/instrumentation-mysql2';
+import { MongoDBInstrumentation } from '@opentelemetry/instrumentation-mongodb';
 import { defineIntegration } from '@sentry/core';
 import type { Integration, IntegrationFn } from '@sentry/types';
 
-import { addOriginToSpan } from '../utils/addOriginToSpan';
+import { addOriginToSpan } from '../../utils/addOriginToSpan';
 import { NodePerformanceIntegration } from './NodePerformanceIntegration';
 
-const _mysql2Integration = (() => {
+const _mongoIntegration = (() => {
   return {
-    name: 'Mysql2',
+    name: 'Mongo',
     setupOnce() {
       registerInstrumentations({
         instrumentations: [
-          new MySQL2Instrumentation({
+          new MongoDBInstrumentation({
             responseHook(span) {
-              addOriginToSpan(span, 'auto.db.otel.mysql2');
+              addOriginToSpan(span, 'auto.db.otel.mongo');
             },
           }),
         ],
@@ -24,20 +24,20 @@ const _mysql2Integration = (() => {
   };
 }) satisfies IntegrationFn;
 
-export const mysql2Integration = defineIntegration(_mysql2Integration);
+export const mongoIntegration = defineIntegration(_mongoIntegration);
 
 /**
- * MySQL2 integration
+ * MongoDB integration
  *
- * Capture tracing data for mysql2
+ * Capture tracing data for MongoDB.
  *
- * @deprecated Use `mysql2Integration()` instead.
+ * @deprecated Use `mongoIntegration()` instead.
  */
-export class Mysql2 extends NodePerformanceIntegration<void> implements Integration {
+export class Mongo extends NodePerformanceIntegration<void> implements Integration {
   /**
    * @inheritDoc
    */
-  public static id: string = 'Mysql2';
+  public static id: string = 'Mongo';
 
   /**
    * @inheritDoc
@@ -47,15 +47,15 @@ export class Mysql2 extends NodePerformanceIntegration<void> implements Integrat
   public constructor() {
     super();
     // eslint-disable-next-line deprecation/deprecation
-    this.name = Mysql2.id;
+    this.name = Mongo.id;
   }
 
   /** @inheritDoc */
   public setupInstrumentation(): void | Instrumentation[] {
     return [
-      new MySQL2Instrumentation({
+      new MongoDBInstrumentation({
         responseHook(span) {
-          addOriginToSpan(span, 'auto.db.otel.mysql2');
+          addOriginToSpan(span, 'auto.db.otel.mongo');
         },
       }),
     ];
