+++ packages/node-experimental/src/integrations/tracing/mysql2.ts
@@ -1,21 +1,21 @@
 import type { Instrumentation } from '@opentelemetry/instrumentation';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
-import { MongooseInstrumentation } from '@opentelemetry/instrumentation-mongoose';
+import { MySQL2Instrumentation } from '@opentelemetry/instrumentation-mysql2';
 import { defineIntegration } from '@sentry/core';
 import type { Integration, IntegrationFn } from '@sentry/types';
 
-import { addOriginToSpan } from '../utils/addOriginToSpan';
+import { addOriginToSpan } from '../../utils/addOriginToSpan';
 import { NodePerformanceIntegration } from './NodePerformanceIntegration';
 
-const _mongooseIntegration = (() => {
+const _mysql2Integration = (() => {
   return {
-    name: 'Mongoose',
+    name: 'Mysql2',
     setupOnce() {
       registerInstrumentations({
         instrumentations: [
-          new MongooseInstrumentation({
+          new MySQL2Instrumentation({
             responseHook(span) {
-              addOriginToSpan(span, 'auto.db.otel.mongoose');
+              addOriginToSpan(span, 'auto.db.otel.mysql2');
             },
           }),
         ],
@@ -24,20 +24,20 @@ const _mongooseIntegration = (() => {
   };
 }) satisfies IntegrationFn;
 
-export const mongooseIntegration = defineIntegration(_mongooseIntegration);
+export const mysql2Integration = defineIntegration(_mysql2Integration);
 
 /**
- * Mongoose integration
+ * MySQL2 integration
  *
- * Capture tracing data for Mongoose.
+ * Capture tracing data for mysql2
  *
- * @deprecated Use `mongooseIntegration()` instead.
+ * @deprecated Use `mysql2Integration()` instead.
  */
-export class Mongoose extends NodePerformanceIntegration<void> implements Integration {
+export class Mysql2 extends NodePerformanceIntegration<void> implements Integration {
   /**
    * @inheritDoc
    */
-  public static id: string = 'Mongoose';
+  public static id: string = 'Mysql2';
 
   /**
    * @inheritDoc
@@ -47,15 +47,15 @@ export class Mongoose extends NodePerformanceIntegration<void> implements Integr
   public constructor() {
     super();
     // eslint-disable-next-line deprecation/deprecation
-    this.name = Mongoose.id;
+    this.name = Mysql2.id;
   }
 
   /** @inheritDoc */
   public setupInstrumentation(): void | Instrumentation[] {
     return [
-      new MongooseInstrumentation({
+      new MySQL2Instrumentation({
         responseHook(span) {
-          addOriginToSpan(span, 'auto.db.otel.mongoose');
+          addOriginToSpan(span, 'auto.db.otel.mysql2');
         },
       }),
     ];
